Move progress dialog from SignupGoogleFragment to GoogleFragment
